Started adding the DSUtils changes
