+++ DS_Map/HeaderSearch.cs
@@ -180,7 +180,7 @@ namespace DSPRE {
 
                     for (ushort i = startID; i < finalID; i++) {
                         MapHeader h;
-                        if (ROMToolboxDialog.flag_DynamicHeadersPatchApplied || ROMToolboxDialog.CheckFilesDynamicHeadersPatchApplied()) {
+                        if (PatchToolboxDialog.flag_DynamicHeadersPatchApplied || PatchToolboxDialog.CheckFilesDynamicHeadersPatchApplied()) {
                             h = MapHeader.LoadFromFile(RomInfo.gameDirs[DirNames.dynamicHeaders].unpackedDir + "\\" + i.ToString("D4"), i, 0);
                         } else {
                             h = MapHeader.LoadFromARM9(i);
