+++ DS_Map/WildEditorDPPt.cs
@@ -22,7 +22,7 @@ namespace DSPRE {
             Dictionary<int, string> EncounterFileLocationName = new Dictionary<int, string>();
 
             for (ushort i = 0; i < totalNumHeaderFiles; i++) {
-                if (ROMToolboxDialog.flag_DynamicHeadersPatchApplied || ROMToolboxDialog.CheckFilesDynamicHeadersPatchApplied()) {
+                if (PatchToolboxDialog.flag_DynamicHeadersPatchApplied || PatchToolboxDialog.CheckFilesDynamicHeadersPatchApplied()) {
                     tempMapHeader = MapHeader.LoadFromFile(RomInfo.gameDirs[DirNames.dynamicHeaders].unpackedDir + "\\" + i.ToString("D4"), i, 0);
                 } else {
                     tempMapHeader = MapHeader.LoadFromARM9(i);
